* libmisc/copydir.c, src/usermod.c, lib/prototypes.h: The uid and   gid parameters can be set to -1 to indicate that the original   owners must be kept. Change the types from uid_t/gid_t to a   long int (signed). * libmisc/copydir.c: Change the copy_entry(), copy_dir(),   copy_symlink(), copy_special(), and copy_file() prototypes   accordingly. * lib/prototypes.h: Add the parameters' name for the   libmisc/copydir.c functions.

+++ src/usermod.c
@@ -1328,8 +1328,8 @@ static void move_home (void)
 					fail_exit (E_HOMEDIR);
 				}
 				if (copy_tree (user_home, user_newhome,
-					       uflg ? user_newid : -1,
-					       gflg ? user_newgid : -1) == 0) {
+					       uflg ? (long int)user_newid : -1,
+					       gflg ? (long int)user_newgid : -1) == 0) {
 					if (remove_tree (user_home) != 0 ||
 					    rmdir (user_home) != 0)
 						fprintf (stderr,
